+++ examples/curve.rs
@@ -1,7 +1,7 @@
 // curve.rs
 extern crate footile;
 
-use footile::{Gray8,PathBuilder,Plotter};
+use footile::{PathBuilder,Plotter};
 
 fn main() -> Result<(), std::io::Error> {
     let path = PathBuilder::new().relative().pen_width(0.0)
@@ -9,6 +9,6 @@ fn main() -> Result<(), std::io::Error> {
                            .pen_width(18.0)
                            .cubic_to(-64.0, -48.0, -64.0, 80.0, 0.0, 32.0)
                            .build();
-    let mut p = Plotter::<Gray8>::new(128, 128);
+    let mut p = Plotter::new(128, 128);
     p.stroke(&path).write_png("./curve.png")
 }