+++ examples/figure.rs
@@ -1,7 +1,7 @@
 // figure.rs
 extern crate footile;
 
-use footile::{Gray8,FillRule,PathBuilder,Plotter};
+use footile::{FillRule,PathBuilder,Plotter};
 
 fn main() -> Result<(), std::io::Error> {
     let path = PathBuilder::new().relative()
@@ -14,7 +14,6 @@ fn main() -> Result<(), std::io::Error> {
                            .line_to(-28.0, 4.0)
                            .line_to(12.0, -28.0)
                            .close().build();
-    let mut p = Plotter::<Gray8>::new(64, 64);
-    p.fill(&path, FillRule::NonZero);
-    p.write_png("./figure.png")
+    let mut p = Plotter::new(64, 64);
+    p.fill(&path, FillRule::NonZero).write_png("./figure.png")
 }