+++ examples/fishy.rs
@@ -1,7 +1,7 @@
 // fishy.rs
 extern crate footile;
 
-use footile::{FillRule,PathBuilder,Plotter,Rgba8};
+use footile::{FillRule,PathBuilder,Plotter,Raster,Rgba8};
 
 fn main() -> Result<(), std::io::Error> {
     let fish = PathBuilder::new().relative().pen_width(3.0)
@@ -18,8 +18,9 @@ fn main() -> Result<(), std::io::Error> {
                           .line_to(-8.0, 8.0)
                           .build();
     let mut p = Plotter::new(128, 128);
-    p.fill(&fish, FillRule::NonZero).over(Rgba8::rgb(127, 96, 96));
-    p.stroke(&fish).over(Rgba8::rgb(255, 208, 208));
-    p.stroke(&eye).over(Rgba8::rgb(0, 0, 0));
-    p.write_png("./fishy.png")
+    let mut r = Raster::new(p.width(), p.height());
+    r.over(p.fill(&fish, FillRule::NonZero), Rgba8::rgb(127, 96, 96));
+    r.over(p.stroke(&fish), Rgba8::rgb(255, 208, 208));
+    r.over(p.stroke(&eye), Rgba8::rgb(0, 0, 0));
+    r.write_png("./fishy.png")
 }