+++ examples/over.rs
@@ -1,7 +1,7 @@
 // over.rs
 extern crate footile;
 
-use footile::{Gray8,PathBuilder,Plotter};
+use footile::{PathBuilder,Plotter};
 
 fn main() -> Result<(), std::io::Error> {
     let path = PathBuilder::new().relative().pen_width(8.0)
@@ -12,6 +12,6 @@ fn main() -> Result<(), std::io::Error> {
                            .line_to(16.0, -16.0)
                            .line_to(0.0, 32.0)
                            .build();
-    let mut p = Plotter::<Gray8>::new(64, 64);
+    let mut p = Plotter::new(64, 64);
     p.stroke(&path).write_png("./over.png")
 }